Update pb.ArmPosition to use floats for all values

+++ arm/arm.go
@@ -32,9 +32,9 @@ type Arm interface {
 // See kinmath/axisAngle.go for a math explnation
 func NewPositionFromMetersAndAngleAxis(x, y, z, th, rx, ry, rz float64) *pb.ArmPosition {
 	return &pb.ArmPosition{
-		X:     int64(x * 1000),
-		Y:     int64(y * 1000),
-		Z:     int64(z * 1000),
+		X:     x * 1000,
+		Y:     y * 1000,
+		Z:     z * 1000,
 		RX:    rx,
 		RY:    ry,
 		RZ:    rz,
